requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.5
django-cors-headers==3.1.1
djangorestframework==3.10.3
gunicorn==19.9.0
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
static3==0.7.0
whitenoise==4.1.4
